# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: accounts
    description: Player accounts (one per identity)
  - name: characters
    description: Per-account characters and saves
  - name: prefabs
    description: Scriptable entity definitions (monsters, NPCs, items, vfx, ...)
  - name: spawn-rules
    description: Where prefabs may appear in the world
  - name: asset-studio
    description: Read-only browse + tagging of every R2 asset

paths:
  /assets/studio/catalog:
    get:
      operationId: getStudioCatalog
      tags: [asset-studio]
      summary: List every R2 asset, grouped by file kind
      description: |
        Read-through listing of every object in the primary R2 assets bucket,
        classified by file extension into 3D models, textures, VFX, audio,
        and other. Returns the public URL for each asset (built from
        `OBJECT_STORAGE_PUBLIC_URL` / `OBJECT_STORAGE_PUBLIC_R2_URL`) so the
        frontend can preview directly without round-tripping the API server.
      responses:
        "200":
          description: The full asset catalog
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StudioCatalog" }
        "502":
          description: R2 listing failed

  /assets/studio/tags:
    get:
      operationId: listStudioTags
      tags: [asset-studio]
      summary: List every saved asset tag
      description: |
        Returns the full map of structured tags (gear slot, character form,
        grudge UUID, notes) keyed by R2 object key. Used by the Asset Studio
        on load to hydrate its in-memory store from the cloud.
      responses:
        "200":
          description: Tag map
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StudioTagMap" }
        "503":
          description: D1 catalog unavailable
    delete:
      operationId: clearStudioTags
      tags: [asset-studio]
      summary: Delete every saved asset tag
      responses:
        "204": { description: Cleared }
        "503": { description: D1 catalog unavailable }

  /assets/studio/tags/{key}:
    parameters:
      - name: key
        in: path
        required: true
        description: URL-encoded R2 object key.
        schema: { type: string }
    put:
      operationId: upsertStudioTag
      tags: [asset-studio]
      summary: Save or clear a single asset tag
      description: |
        Upserts the structured tag for one asset. If every field is empty or
        whitespace, the row is deleted and a 204 is returned instead of a 200.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/StudioTagInput" }
      responses:
        "200":
          description: Saved tag
          content:
            application/json:
              schema: { $ref: "#/components/schemas/StudioTagUpsertResult" }
        "204": { description: Tag cleared (all fields empty) }
        "400": { description: Invalid body or empty key }
        "503": { description: D1 catalog unavailable }
    delete:
      operationId: deleteStudioTag
      tags: [asset-studio]
      summary: Delete a single asset tag
      responses:
        "204": { description: Deleted }
        "503": { description: D1 catalog unavailable }

  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema: { $ref: "#/components/schemas/HealthStatus" }

  /accounts/upsert:
    post:
      operationId: upsertAccount
      tags: [accounts]
      summary: Get-or-create an account by grudge id
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpsertAccountRequest" }
      responses:
        "200":
          description: Account record
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }

  /accounts/{grudgeId}:
    get:
      operationId: getAccountByGrudgeId
      tags: [accounts]
      parameters:
        - { name: grudgeId, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Account" }
        "404": { description: Not found }

  /characters:
    get:
      operationId: listCharacters
      tags: [characters]
      parameters:
        - { name: accountId, in: query, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: List of characters under an account
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Character" }
    post:
      operationId: createCharacter
      tags: [characters]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCharacterRequest" }
      responses:
        "201":
          description: Created character
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Character" }

  /characters/{id}:
    get:
      operationId: getCharacter
      tags: [characters]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Character
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Character" }
        "404": { description: Not found }
    put:
      operationId: updateCharacter
      tags: [characters]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCharacterRequest" }
      responses:
        "200":
          description: Updated character
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Character" }
    delete:
      operationId: deleteCharacter
      tags: [characters]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Deleted }

  /prefabs:
    get:
      operationId: listPrefabs
      tags: [prefabs]
      parameters:
        - { name: kind, in: query, required: false, schema: { $ref: "#/components/schemas/PrefabKind" } }
        - { name: includeDrafts, in: query, required: false, schema: { type: boolean, default: false } }
      responses:
        "200":
          description: Prefab list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Prefab" }
    post:
      operationId: createPrefab
      tags: [prefabs]
      security: [{ adminBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpsertPrefabRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Prefab" }
        "401": { description: Admin token required }

  /prefabs/{id}:
    get:
      operationId: getPrefab
      tags: [prefabs]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Prefab
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Prefab" }
        "404": { description: Not found }
    put:
      operationId: updatePrefab
      tags: [prefabs]
      security: [{ adminBearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpsertPrefabRequest" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Prefab" }
        "401": { description: Admin token required }
    delete:
      operationId: deletePrefab
      tags: [prefabs]
      security: [{ adminBearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Deleted }
        "401": { description: Admin token required }

  /admin/players/search:
    get:
      operationId: searchPlayers
      tags: [accounts]
      security: [{ adminBearer: [] }]
      description: |
        Search for accounts by substring match on grudge id, display name, or
        any of their character names. Each result includes the matching account
        plus all of its characters. Admin only.
      parameters:
        - { name: q, in: query, required: true, schema: { type: string, minLength: 1, maxLength: 128 } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PlayerSearchResponse" }
        "400": { description: Missing or invalid query }
        "401": { description: Admin token required }

  /spawn-rules:
    get:
      operationId: listSpawnRules
      tags: [spawn-rules]
      parameters:
        - { name: prefabId, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: Spawn rules
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SpawnRule" }
    post:
      operationId: createSpawnRule
      tags: [spawn-rules]
      security: [{ adminBearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSpawnRuleRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SpawnRule" }
        "401": { description: Admin token required }

  /spawn-rules/{id}:
    delete:
      operationId: deleteSpawnRule
      tags: [spawn-rules]
      security: [{ adminBearer: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "204": { description: Deleted }
        "401": { description: Admin token required }

components:
  securitySchemes:
    adminBearer:
      type: http
      scheme: bearer
      bearerFormat: token

  schemas:
    HealthStatus:
      type: object
      properties:
        status: { type: string }
      required: [status]

    PrefabKind:
      type: string
      enum: [monster, npc, player_body, item, vfx, container, structure]

    Account:
      type: object
      properties:
        id: { type: string, format: uuid }
        grudgeId: { type: string }
        displayName: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, grudgeId, createdAt, updatedAt]

    UpsertAccountRequest:
      type: object
      properties:
        grudgeId: { type: string }
        displayName: { type: string, nullable: true }
      required: [grudgeId]

    Character:
      type: object
      properties:
        id: { type: string, format: uuid }
        accountId: { type: string, format: uuid }
        name: { type: string }
        config: {}
        saveData: { nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        lastPlayedAt: { type: string, format: date-time, nullable: true }
      required: [id, accountId, name, config, createdAt, updatedAt]

    CreateCharacterRequest:
      type: object
      properties:
        accountId: { type: string, format: uuid }
        name: { type: string }
        config: {}
      required: [accountId, name, config]

    UpdateCharacterRequest:
      type: object
      properties:
        name: { type: string }
        config: {}
        saveData: {}
        touchLastPlayed: { type: boolean }

    PlayerSearchMatch:
      type: object
      properties:
        account: { $ref: "#/components/schemas/Account" }
        characters:
          type: array
          items: { $ref: "#/components/schemas/Character" }
        matchedFields:
          type: array
          description: Which fields matched the query for this account.
          items:
            type: string
            enum: [grudgeId, displayName, characterName]
      required: [account, characters, matchedFields]

    PlayerSearchResponse:
      type: object
      properties:
        q: { type: string }
        returnedCount:
          type: integer
          description: Number of entries actually returned in `results`.
        hasMore:
          type: boolean
          description: |
            True when at least one matching account was elided by the limit
            (or by the per-source candidate cap). Use this instead of an
            exact total — the server does not run an unbounded count.
        results:
          type: array
          items: { $ref: "#/components/schemas/PlayerSearchMatch" }
      required: [q, returnedCount, hasMore, results]

    Prefab:
      type: object
      properties:
        id: { type: string }
        kind: { $ref: "#/components/schemas/PrefabKind" }
        name: { type: string }
        description: { type: string, nullable: true }
        modelPath: { type: string, nullable: true }
        texturePath: { type: string, nullable: true }
        scale: { type: number }
        data: {}
        tags:
          type: array
          items: { type: string }
        draft: { type: boolean }
        version: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, kind, name, scale, data, tags, draft, version, createdAt, updatedAt]

    UpsertPrefabRequest:
      type: object
      properties:
        id: { type: string }
        kind: { $ref: "#/components/schemas/PrefabKind" }
        name: { type: string }
        description: { type: string, nullable: true }
        modelPath: { type: string, nullable: true }
        texturePath: { type: string, nullable: true }
        scale: { type: number }
        data: {}
        tags:
          type: array
          items: { type: string }
        draft: { type: boolean }
      required: [id, kind, name]

    SpawnRule:
      type: object
      properties:
        id: { type: string, format: uuid }
        prefabId: { type: string }
        biome: { type: string, nullable: true }
        minDistance: { type: number }
        maxDistance: { type: number }
        minWave: { type: integer }
        weight: { type: integer }
        enabled: { type: boolean }
        createdAt: { type: string, format: date-time }
      required: [id, prefabId, minDistance, maxDistance, minWave, weight, enabled, createdAt]

    StudioAssetKind:
      type: string
      description: |
        Coarse classification used to group assets in the Asset Studio gallery.
      enum: [model, texture, vfx, audio, other]

    StudioAsset:
      type: object
      properties:
        key:          { type: string, description: "Object key in the R2 bucket" }
        filename:     { type: string, description: "Last path segment of the key" }
        ext:          { type: string, description: "Lowercase file extension without the dot" }
        size:         { type: integer, description: "Object size in bytes" }
        lastModified: { type: string, description: "ISO timestamp from R2" }
        contentType:  { type: string, description: "Best-effort MIME guess from extension" }
        publicUrl:    { type: string, nullable: true, description: "Public CDN URL, or null if no public base is configured" }
        kind:         { $ref: "#/components/schemas/StudioAssetKind" }
      required: [key, filename, ext, size, lastModified, contentType, publicUrl, kind]

    StudioGroup:
      type: object
      properties:
        kind:       { $ref: "#/components/schemas/StudioAssetKind" }
        label:      { type: string, description: "Human-readable section title" }
        count:      { type: integer }
        totalBytes: { type: integer }
        assets:
          type: array
          items: { $ref: "#/components/schemas/StudioAsset" }
      required: [kind, label, count, totalBytes, assets]

    StudioCatalog:
      type: object
      properties:
        generatedAt:   { type: string, description: "ISO timestamp the catalog was assembled" }
        bucket:        { type: string, description: "R2 bucket the catalog was read from" }
        publicUrlBase: { type: string, nullable: true, description: "Base URL used to construct publicUrl values" }
        totalCount:    { type: integer }
        totalBytes:    { type: integer }
        truncated:     { type: boolean, description: "True if the listing hit the safety cap and is incomplete" }
        groups:
          type: array
          items: { $ref: "#/components/schemas/StudioGroup" }
      required: [generatedAt, bucket, publicUrlBase, totalCount, totalBytes, truncated, groups]

    StudioTag:
      type: object
      description: Structured tag attached to a single R2 asset.
      properties:
        gearSlot:      { type: string }
        characterForm: { type: string }
        grudgeUuid:    { type: string }
        notes:         { type: string }
        updatedAt:     { type: string, description: "ISO timestamp of the last save" }
      required: [updatedAt]

    StudioTagInput:
      type: object
      description: |
        Editable fields for a studio tag. Every field is optional; sending
        them all empty or null clears the tag entirely.
      properties:
        gearSlot:      { type: string, nullable: true }
        characterForm: { type: string, nullable: true }
        grudgeUuid:    { type: string, nullable: true }
        notes:         { type: string, nullable: true }

    StudioTagMap:
      type: object
      properties:
        tags:
          type: object
          description: Map keyed by raw (un-encoded) R2 object key.
          additionalProperties: { $ref: "#/components/schemas/StudioTag" }
        count: { type: integer }
      required: [tags, count]

    StudioTagUpsertResult:
      type: object
      properties:
        key: { type: string }
        tag: { $ref: "#/components/schemas/StudioTag" }
      required: [key, tag]

    CreateSpawnRuleRequest:
      type: object
      properties:
        prefabId: { type: string }
        biome: { type: string, nullable: true }
        minDistance: { type: number }
        maxDistance: { type: number }
        minWave: { type: integer }
        weight: { type: integer }
        enabled: { type: boolean }
      required: [prefabId]